requirements.txt
appdirs==1.4.3
argh==0.26.2
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
bootstrap4==0.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.3
defusedxml==0.5.0
distlib==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.5
Django-Bootstrap3-Validator==0.3.3
django-crispy-forms==1.9.0
django-google-maps==0.12.1
django-heroku==0.3.1
django-leaderboard==0.1.0
django-model-utils==3.1.2
django-quiz-app==0.5.1
django-stubs==1.5.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
docopt==0.6.2
filelock==3.0.10
gitignore==0.0.8
gunicorn==20.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
oauthlib==3.0.1
pathtools==0.1.2
pillow==7.0.0
pluggy==0.9.0
psycopg2==2.8.5
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
PyMySQL==0.9.3
pytz==2018.9
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
virtualenv==20.0.7
whitenoise==5.0.1
wrapt==1.11.2
